Adding Functionality to the Buttons
Adding the functionality of clicking the buttons to using the functions of ".add" and ".remove" to remove classes which allow the overlay popup to show when buttons were pressed.

Also simplified the code by creating variables that contained the functions instead to add and remove the classes.

This overall made the code more declarative and concise.