Completed Modal Window Project
In this update I add the functionality of pressing the "esc" on the keyboard to exit the modal windows.

I did this using an if statement which was originally two if statements but simplified down into one using "&&" as seen in the final code.

This exercise was a lot of fun and I look forward to doing more with key presses like this. The possibilities are truly endless!